chore: rename buttonBlob to ButtonLinkBlob
